added ratelimiting and secret for upload

+++ routes/routes.js
@@ -6,6 +6,7 @@ import {config} from '../controller/config.js'
 
 const router = express.Router()
 const uploads = config.uploadFolder
+const uploadSecretValue = config.uploadSecretValue
 
 //Multer custom storage for a proper file name
 const storage = multer.diskStorage({
@@ -27,13 +28,19 @@ const upload = multer({ storage: storage })
 //upload.single: Accepts a single file with the name "fieldname". The single file will be stored in req.file.
 router.post('/image', upload.single('image'), async function(req,res) {
     try {
+
+        const uploadSecret = req.body['upload-secret']
+        if (!uploadSecret || (uploadSecret !== uploadSecretValue)){
+            return res.status(401).send("Not authorized")
+        }
+
         const filePath = req.file.path
         const buffer = await fs.readFile(filePath)
         const type = await fileTypeFromBuffer(buffer) // Check file type
 
         if (!type || !['image/jpeg', 'image/png', 'image/bmp'].includes(type.mime)) {
-            await fs.unlink(filePath); // Delete invalid file
-            return res.status(400).send('Invalid file type');
+            await fs.unlink(filePath) // Delete invalid file
+            return res.status(400).send('Invalid file type')
         }
 
         // Return the file name to the user
@@ -48,7 +55,7 @@ router.post('/image', upload.single('image'), async function(req,res) {
 
     } catch (error) {
         console.log(error)
-        res.status(500).send("An error occurred during file upload")
+        return res.status(500).send("An error occurred during file upload")
     }
 })
